// backend/src/modules/bulkmapping/bm.enum.ts
export enum SpsSortColumn {
    GRADE = 'GRADE',
    JOB_TITLE = 'JOB_TITLE',
    FUNCTION = 'FUNCTION',
    FUNCTIONS = 'FUNCTIONS',
    LEVELS = 'LEVELS',
    LEVEL = 'LEVEL',
    MODIFIED_ON = 'MODIFIED_ON',
}

export enum SpsFilters{
    LEVELS = 'LEVELS',
    GRADES = 'GRADES',
    CUSTOMGRADES = 'CUSTOMGRADESET',
    FUNCTIONS = 'FUNCTIONS',
    SUBFUNCTIONS = 'SUBFUNCTIONS',
}

export enum SortBy {
    ASC = 'ASC',
    DESC = 'DESC',
}

export enum MetadataFilterValues {
    GRADES = 'GRADES',
    LEVELS = 'LEVELS',
    FUNCTIONS = 'FUNCTIONS',
    CUSTOMGRADESET = 'CUSTOMGRADESET',
}

export enum MetadataFilterValuesRemaped {
    LEVELS = 'Levels',
    FUNCTIONS = 'Functions',
    GRADES = 'Grades',
}

export enum GradeIndication {
    GRADES = 'GRADES',
    CUSTOM_GRADES = 'CUSTOM_GRADES',
}

export enum MssqlDataSets {
    SPS_DATA_RAW = 'DATA',
    SPS_IDS_RAW = 'ID',
    DATA_RAW = 'DATA',
    IDS_RAW = 'ID',
}

export enum CompetenceModelEntityType {
    FACTOR = 'FACTOR',
    CLUSTER = 'CLUSTER',
    COMPETENCY = 'COMPETENCY',
}